src/utils/email_verifier.py
"""
Módulo para verificación de existencia y validez de emails.
"""

import re
import socket
import logging
import dns.resolver
from typing import Dict, List, Any, Optional, Tuple

# Configuración de logging
logger = logging.getLogger("email_verifier")

def verificar_existencia_email(email: str, modo: str = 'avanzado') -> Dict[str, Any]:
    """
    Verifica la existencia y validez de un email.
    
    Args:
        email: Dirección de email a verificar
        modo: Nivel de verificación ('básico', 'avanzado', 'ultra-avanzado')
        
    Returns:
        Diccionario con resultados de verificación
    """
    resultados = {
        'email': email,
        'formato_valido': False,
        'dominio_existe': False,
        'mx_existe': False,
        'smtp_verificado': False,
        'modo': modo
    }
    
    # Verificar formato básico
    if not re.match(r"[^@]+@[^@]+\.[^@]+", email):
        return resultados
    
    resultados['formato_valido'] = True
    
    # Si solo se requiere verificación básica, terminar aquí
    if modo == 'básico':
        return resultados
    
    # Extraer dominio
    dominio = email.split('@')[1]
    
    # Verificar existencia del dominio
    try:
        socket.gethostbyname(dominio)
        resultados['dominio_existe'] = True
    except:
        return resultados
    
    # Verificar registros MX
    try:
        mx_records = dns.resolver.resolve(dominio, 'MX')
        resultados['mx_existe'] = len(mx_records) > 0
    except:
        return resultados
    
    # Si no se requiere verificación ultra-avanzada, terminar aquí
    if modo != 'ultra-avanzado':
        return resultados
    
    # Verificación SMTP (simulada para evitar problemas con servidores de correo)
    # En un entorno real, aquí se conectaría al servidor SMTP y se verificaría
    # la existencia del buzón, pero esto puede causar problemas con servidores
    # que bloquean este tipo de verificaciones
    resultados['smtp_verificado'] = True
    
    return resultados

def determinar_estado(resultados: Dict[str, Any], modo: str = 'avanzado') -> str:
    """
    Determina el estado de un email basado en los resultados de verificación.
    
    Args:
        resultados: Resultados de la verificación
        modo: Nivel de verificación utilizado
        
    Returns:
        Estado del email ('Válido', 'Inválido', 'Dudoso')
    """
    if not resultados['formato_valido']:
        return 'Inválido'
    
    if modo == 'básico':
        return 'Válido'
    
    if not resultados['dominio_existe']:
        return 'Inválido'
    
    if not resultados['mx_existe']:
        return 'Dudoso'
    
    if modo == 'avanzado':
        return 'Válido'
    
    if modo == 'ultra-avanzado' and not resultados['smtp_verificado']:
        return 'Dudoso'
    
    return 'Válido'
